Initial commit
Waited until day18 to create a Github repo and push my work for public
ridicule.  I went back and cleaned up a few examples to make sure they
run with unmodified input and in a few cases to clarify the output so
it's more apparent what value is the answer.

Signed-off-by: Charles Steinkuehler <cstein@newtek.com>